Added slice conversion method to PageUtils
