+++ src/main/java/com/data/utils/page/PageUtils.java
@@ -1,16 +1,15 @@
 package com.data.utils.page;
 
-import java.lang.reflect.Field;
 import java.util.ArrayList;
-import java.util.Comparator;
 import java.util.List;
 import java.util.Map;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Slice;
+import org.springframework.data.domain.SliceImpl;
 import org.springframework.data.domain.Sort;
-import org.springframework.data.domain.Sort.Direction;
 import org.springframework.data.domain.Sort.Order;
 import org.springframework.util.CollectionUtils;
 
@@ -45,35 +44,43 @@ public class PageUtils {
     }
 
     int start = Math.toIntExact(pageable.getOffset());
-
     int end = Math.min((start + pageable.getPageSize()), list.size());
 
-    if (start >= list.size()) {
-      return Page.empty();
-    }
+    if (start >= list.size()) return Page.empty();
 
-    List<T> sortedList = sortList(list, pageable);
+    List<T> sortedList = SortingUtils.sortList(list, pageable);
 
     return new PageImpl<>(sortedList.subList(start, end), pageable, sortedList.size());
   }
 
   /**
-   * Sorts the given list of objects using the sort criteria from the given {@link Pageable} object.
+   * Converts the given list of objects to a {@link Slice} object, using the given {@link Pageable}
+   * object for pagination and sorting.
    *
-   * @param list the list of objects to sort
-   * @param pageable the {@link Pageable} object containing the sort criteria
+   * @param list the list of objects to include in the slice
+   * @param pageable the {@link Pageable} object containing the pagination and sort criteria
+   * @return a {@link Slice} object with the specified page size, page number, and sorting criteria
    */
-  public static <T> List<T> sortList(List<T> list, Pageable pageable) {
+  public static <T> Slice<T> toSlice(List<T> list, Pageable pageable) {
 
-    List<T> copy = new ArrayList<>(list);
+    if (CollectionUtils.isEmpty(list)) {
+      return new SliceImpl<>(List.of(), pageable, false);
+    }
 
-    List<Sort.Order> orders = pageable.getSort().get().toList();
+    if (pageable == null) {
+      pageable = Pageable.ofSize(list.size());
+    }
 
-    Comparator<T> comparator = createComparator(orders);
+    int start = Math.toIntExact(pageable.getOffset());
+    int end = Math.min((start + pageable.getPageSize()), list.size());
 
-    copy.sort(comparator);
+    if (start >= list.size()) return Page.empty();
 
-    return copy;
+    List<T> sortedList = SortingUtils.sortList(list, pageable);
+
+    boolean hasNext = end < sortedList.size();
+
+    return new SliceImpl<>(sortedList.subList(start, end), pageable, hasNext);
   }
 
   /**
@@ -102,85 +109,5 @@ public class PageUtils {
       }
     }
     return PageRequest.of(pageable.getPageNumber(), pageable.getPageSize(), Sort.by(orders));
-  }
-
-  /**
-   * Creates a Comparator for sorting a list of objects using the given sort orders and class.
-   *
-   * @param orders the sort orders to use for sorting
-   * @param <T> the type of the objects in the list
-   * @return A Comparator that can be used to sort the list of objects
-   */
-  @SuppressWarnings("unchecked")
-  public static <T> Comparator<T> createComparator(List<Sort.Order> orders) {
-    return (o1, o2) -> {
-      for (Sort.Order order : orders) {
-        String sortField = order.getProperty();
-        Sort.Direction direction = order.getDirection();
-
-        Object value1 = getFlattenedValue(o1, sortField);
-        Object value2 = getFlattenedValue(o2, sortField);
-
-        if (value1 == null && value2 == null) {
-          continue;
-        } else if (value1 == null) {
-          return (direction == Direction.ASC) ? -1 : 1;
-        } else if (value2 == null) {
-          return (direction == Direction.ASC) ? 1 : -1;
-        }
-        if (value1 instanceof String stringValue1 && value2 instanceof String stringValue2) {
-          int result = stringValue1.compareToIgnoreCase(stringValue2);
-          return (direction == Direction.ASC) ? result : -result;
-        }
-        if (value1 instanceof Comparable<?> && value2 instanceof Comparable<?>) {
-          int result = ((Comparable<Object>) value1).compareTo(value2);
-          return (direction == Direction.ASC) ? result : -result;
-        }
-      }
-      return 0;
-    };
-  }
-
-  private static Object getFlattenedValue(Object obj, String sortField) {
-    String[] keys = sortField.split("_");
-    Object currentObj = obj;
-
-    for (String key : keys) {
-      if (currentObj == null) {
-        break;
-      }
-
-      if (currentObj instanceof Map<?, ?>) {
-        currentObj = ((Map<?, ?>) currentObj).get(key);
-      } else {
-        currentObj = getFieldOrNestedField(currentObj, key);
-      }
-    }
-
-    return currentObj;
-  }
-
-  private static Object getFieldOrNestedField(Object obj, String fieldName) {
-    Field field = getField(obj.getClass(), fieldName);
-    if (field != null) {
-      field.setAccessible(true);
-      try {
-        return field.get(obj);
-      } catch (IllegalAccessException e) {
-        throw new IllegalArgumentException("Error accessing field: " + fieldName, e);
-      }
-    }
-    return null;
-  }
-
-  private static Field getField(Class<?> clazz, String fieldName) {
-    try {
-      return clazz.getDeclaredField(fieldName);
-    } catch (NoSuchFieldException e) {
-      if (clazz.getSuperclass() != null) {
-        return getField(clazz.getSuperclass(), fieldName);
-      }
-      return null;
-    }
   }
 }
